ALBOUS DOUBOULOUMBELDOLORES

+++ contracts/CipaCoin.sol
@@ -221,8 +221,10 @@ contract CipaCoin {
 
         require(!alreadyExists, "Le club existe deja.");
 
-        require(getTimesPres(president)<5,
-                "L'etudiant est deja president de 5 clubs, il ne peut plus devenir president d'un nouveau club." );
+        require(
+            getTimesPres(president) < 5,
+            "L'etudiant est deja president de 5 clubs, il ne peut plus devenir president d'un nouveau club."
+        );
 
         clubs.push(
             Club({
@@ -255,11 +257,13 @@ contract CipaCoin {
         students[msg.sender].cipaStudentBalance = 0;
     }
 
-    function getTimesPres(address student) public view returns (uint){
-        uint nbClub = 0;
+    function getTimesPres(address student) public view returns (uint256) {
+        uint256 nbClub = 0;
 
         for (uint256 i = 0; i < clubs.length; i++) {
-            if(getClubPres(i)==student){nbClub++;}
+            if (getClubPres(i) == student) {
+                nbClub++;
+            }
         }
 
         return nbClub;
@@ -283,8 +287,10 @@ contract CipaCoin {
             "L'etudiant est deja president de ce club."
         );
 
-        require(getTimesPres(newPres)<5,
-        "L'etudiant est deja president de 5 clubs, il ne peut plus devenir president d'un nouveau club.");
+        require(
+            getTimesPres(newPres) < 5,
+            "L'etudiant est deja president de 5 clubs, il ne peut plus devenir president d'un nouveau club."
+        );
 
         clubs[clubInt].president = newPres;
         clubs[clubInt].totalCipaOwnedSinceNomination = 0;
